Hero and Sidebar
In this commit I've finished the Hero and Sidebar sections. I gave the Hero section some state called isOpen / setIsOpen using the useState hook.

I then created a toggle function that uses 'setIsOpen' to toggle the 'isOpen' between true and false.

When set to true the sidebar container 's right property is set to 0, and it's -1000px (off screen) when set to false. This is done through a ternary operator within the 'SidebarContainer' styled-component.

I passed the toggle function from the Hero component down to both the Navbar and Sidebar components. I passed the isOpen state down to the Sidebar component.

I gave the NavIcon and Sidebar Icon onClick props that trigger the toggle function and change the isOpen state, thus making the sidebar appear / disappear.

So clicking the pizza icon opens the sidebar and clicking the cross icon on the sidebar closes it.