Proucts, Feature and Footer
In this commit I've finished the Products, Feature and Footer sections.

For the products section I created an array of objects in a seperate data file. I imported this data file in the App.js file, and passed this data down to the Products component.

In this Products component I then used the map() function to create a product card, containing all the relevant info, for each object (product) in the data array.  The styling for each element was again done in styled-components file (productElements) and imported to the Product component.

I then created the Feature component, showcasing a featured pizza.

After the Feature component, I then placed another Product component (in the App.js file.) This time I passed it down a different array of objects from the data file, containing the relevant information for the desserts section. Again the Products section mapped each object, creating all of the Product cards and displaying them like before.

This was a great re-use of the Products component, as it made creating the desserts section very quick and straight forward.

After the second Products component I made the Footer component. This was done using styled-components once again, utilizing react-icons for the social media logos.