Navbar Update
I've given the navbar a dark background and fixed it to the top of the page. I did this because I wanted the pizza icon to be available to click when scrolling through the page.

For the mobile version I have hidden the title and have left only the pizza icon visible, moving it to the middle of the nav bar. I feel this looks better on small screens as it de-clutters the page, making for a cleaner design.